+++ cicada-go/cmd/cicada/main.go
@@ -47,6 +47,15 @@ func main() {
 			fmt.Fprintln(os.Stderr, err)
 			os.Exit(1)
 		}
+	case "connector":
+		if len(os.Args) < 3 || os.Args[2] != "telegram" {
+			fmt.Fprintln(os.Stderr, "usage: cicada connector telegram [options]")
+			os.Exit(2)
+		}
+		if err := runTelegramConnector(os.Args[3:]); err != nil {
+			fmt.Fprintln(os.Stderr, err)
+			os.Exit(1)
+		}
 	default:
 		usage()
 		os.Exit(2)
@@ -238,7 +247,7 @@ func requestJSON(url, method string, body any) error {
 }
 
 func usage() {
-	fmt.Fprintln(os.Stderr, "usage: cicada serve|goal|machine|worker|thread|external|version")
+	fmt.Fprintln(os.Stderr, "usage: cicada serve|goal|machine|worker|thread|external|connector|version")
 }
 
 func envOr(name, fallback string) string {